requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pypi_0
antlr-python-runtime=4.9.2=py38h578d9bd_0
appdirs=1.4.4=pyh9f0ad1d_0
astunparse=1.6.3=pypi_0
audioread=2.1.9=py38h578d9bd_0
blas=1.0=mkl
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.5.30=ha878542_0
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py38h578d9bd_0
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
cloudpickle=1.6.0=pypi_0
coloredlogs=15.0.1=pypi_0
cryptography=3.4.7=py38ha5dfef3_0
cycler=0.10.0=py_2
cython=0.29.21=py38he6710b0_0
decorator=4.4.2=pypi_0
ffmpeg=4.3.1=hca11adc_2
flatbuffers=1.12=pypi_0
freetype=2.10.4=h0708190_1
gast=0.3.3=pypi_0
gettext=0.19.8.1=h0b5b191_1005
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.33.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
humanfriendly=9.2=pypi_0
idna=3.1=pyhd3deb0d_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.4=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
jbig=2.1=h7f98852_2003
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
kapre=0.3.5=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py38h1fd1430_1
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=2.2.1=h9c3ff4c_0
libdeflate=1.7=h7f98852_5
libffi=3.3=he6710b0_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.3.0=h5101ec6_17
libllvm10=10.0.1=he513fc3_3
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.3.0=hf544144_1
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.0=h7f98852_2
llvmlite=0.36.0=py38h4630a5e_0
lz4-c=1.9.3=h9c3ff4c_0
markdown=3.3.4=pypi_0
matplotlib-base=3.4.2=py38hcc49a3a_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
moviepy=1.0.3=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.6=he412f7d_0
networkx=2.5.1=pypi_0
numba=0.53.1=py38h8b71fd7_1
numpy=1.19.5=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
omegaconf=2.1.0=py38h578d9bd_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openl3=0.4.0=pypi_0
openssl=1.1.1k=h7f98852_0
opt-einsum=3.3.0=pypi_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.1.3=py38he6710b0_0
pillow=8.3.1=py38h8e6f84c_0
pip=21.1.3=py38h06a4308_0
pooch=1.4.0=pyhd8ed1ab_0
proglog=0.1.9=pypi_0
protobuf=3.17.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pynndescent=0.5.4=pyh6c4a22f_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h578d9bd_3
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2020.1=py_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=py38h497a2fe_0
randomcolor=0.4.4.5=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pypi_0
resampy=0.2.2=py_0
rsa=4.7.2=pypi_0
scikit-image=0.14.5=pypi_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.6.2=py38had2a1c9_1
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tbb=2020.2=h4bd325d_4
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.4.2=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.2.0=pyh8a188c0_0
tk=8.6.10=hbc83047_0
tornado=6.1=py38h497a2fe_1
tqdm=4.61.2=pypi_0
typing-extensions=3.7.4.3=pypi_0
umap-learn=0.5.1=py38h578d9bd_1
urllib3=1.26.6=pyhd8ed1ab_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
x264=1!161.3030=h7f98852_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zlib=1.2.11=h7b6447c_3
zstd=1.5.0=ha95c52a_0
